[UI] [REFACTO] Relocate loose shared/ files to their proper homes
- shared/TextField.tsx -> components/ui/GenericTextField.tsx (a UI primitive;
  renamed to match its export and avoid the MUI TextField name clash)
- shared/route.ts -> shared/navigation/routes.ts
- shared/darkMenuProps.ts -> theme/darkMenuProps.ts (a theme styling preset)
- Update all importers; src/shared root now holds only subdirectories
- AGENTS.md: refresh the Phase 3 structure backlog

Verified: typecheck, lint, and build all pass.
