+++ src/modules/hub/Hub.tsx
@@ -1,6 +1,6 @@
 import { ProjectExResponseDto } from "@api";
 import { PREDEFINED_PROJECT_TAGS } from "@modules/projects/projectTags";
-import * as urls from "@shared/route";
+import * as urls from "@shared/navigation/routes";
 import { LocalStorageManager } from "@utils/LocalStorageManager";
 
 import { NewHubSection } from "./components/NewHubSection";
