+++ src/modules/projects/ProjectCategoryPage.tsx
@@ -21,7 +21,7 @@ import {
 } from "@modules/projects/projectListUtils";
 import { PREDEFINED_PROJECT_TAGS } from "@modules/projects/projectTags";
 import { useUser } from "@providers/UserProvider";
-import * as urls from "@shared/route";
+import * as urls from "@shared/navigation/routes";
 
 import ProjectCard from "./components/ProjectCard";
