+++ src/modules/projects/components/ProjectCard.tsx
@@ -4,7 +4,7 @@ import {
   ProjectExResponseDto,
   ProjectResponseDto } from "@api";
 import Card from "@modules/projects/components/Card";
-import * as urls from "@shared/route";
+import * as urls from "@shared/navigation/routes";
 import { UserAvatarStack } from "@shared/user/UserProfileLink";
 import { getCachedProjectImageUrl } from "@utils/projectImageCache";
