+++ src/modules/projects/hooks/useCreateProject.ts
@@ -1,5 +1,5 @@
 import { CreateProjectDto, projectControllerCreate } from "@api";
-import * as urls from "@shared/route";
+import * as urls from "@shared/navigation/routes";
 import { LocalStorageManager } from "@utils/LocalStorageManager";
 
 import { useCallback, useState } from "react";
